refactor: Use ref for sessionStats in completeGame
The `sessionStats` object was being captured by the `completeGame` callback. This could lead to outdated stats being used if the `completeGame` function was called after `sessionStats` had been updated but before the callback was executed.

Using a `useRef` to store the latest `sessionStats` ensures that the most up-to-date statistics are always used when finalizing the game.